--- a/src/lexer.cc
+++ b/src/lexer.cc
@@ -5,7 +5,7 @@
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wpedantic"  // allow designated initializers
 
-bool Lexer::Analyse(string& entry, map<string, ReservedWord>& keywords, vector<SynElement>& elements,
+bool Lexer::Analyse(const string& entry, map<string, ReservedWord>& keywords, vector<SynElement>& elements,
                     vector<SynError>& errors) {
     size_t jump;
     for (size_t i = 0; i < entry.size(); i++) {
@@ -52,7 +52,7 @@ void Lexer::Trim(string& s) {
     s.erase(std::find_if(s.rbegin(), s.rend(), [](unsigned char ch) { return !std::isspace(ch); }).base(), s.end());
 }
 
-bool Lexer::ParseString(string& entry, size_t idx, size_t& next_idx, vector<SynError>& errors __attribute__((unused)),
+bool Lexer::ParseString(const string& entry, size_t idx, size_t& next_idx, vector<SynError>& errors __attribute__((unused)),
                         vector<SynElement>& elements) {
     // here we are sure that entry[0] is at least '"'
     for (size_t i = idx + 1; i < entry.size(); i++) {
@@ -69,7 +69,7 @@ bool Lexer::ParseString(string& entry, size_t idx, size_t& next_idx, vector<SynE
     return true;
 }
 
-bool Lexer::ParseSymbol(string& entry, size_t idx, size_t& next_idx, vector<SynError>& errors __attribute__((unused)),
+bool Lexer::ParseSymbol(const string& entry, size_t idx, size_t& next_idx, vector<SynError>& errors __attribute__((unused)),
                         vector<SynElement>& elements) {
     // here we are sure that entry[0] is at least '\''
     for (size_t i = idx + 1; i < entry.size(); i++) {
@@ -84,7 +84,7 @@ bool Lexer::ParseSymbol(string& entry, size_t idx, size_t& next_idx, vector<SynE
     return true;
 }
 
-bool Lexer::ParseProgram(string& entry, size_t idx, size_t& next_idx, vector<SynError>& errors __attribute__((unused)),
+bool Lexer::ParseProgram(const string& entry, size_t idx, size_t& next_idx, vector<SynError>& errors __attribute__((unused)),
                          vector<SynElement>& elements) {
     // here we are sure that entry is at least "<<"
     // find last ">>" or "»"
@@ -111,7 +111,7 @@ bool Lexer::ParseProgram(string& entry, size_t idx, size_t& next_idx, vector<Syn
     return true;
 }
 
-int Lexer::GetBaseAt(string& entry, size_t& next_idx, bool& positive) {
+int Lexer::GetBaseAt(const string& entry, size_t& next_idx, bool& positive) {
     // a regex could be "([+-])?((0[xX])|([0-9][0-9]?[bB]))"
     // regex is not use because dramatically slow
     // entry is scanned from idxStart, searching for [s]abc (sign and 3 first chars)
@@ -154,7 +154,7 @@ int Lexer::GetBaseAt(string& entry, size_t& next_idx, bool& positive) {
     return 10;
 }
 
-bool Lexer::GetNUmberAt(string& entry, size_t idx, size_t& next_idx, int& base, mpreal** r, char delim) {
+bool Lexer::GetNumberAt(const string& entry, size_t idx, size_t& next_idx, int& base, mpreal** r, char delim) {
     stringstream ss;
     string token;
     bool positive = true;
@@ -182,11 +182,11 @@ bool Lexer::GetNUmberAt(string& entry, size_t idx, size_t& next_idx, int& base, 
     return false;
 }
 
-bool Lexer::ParseNumber(string& entry, size_t idx, size_t& next_idx, vector<SynError>& errors,
+bool Lexer::ParseNumber(const string& entry, size_t idx, size_t& next_idx, vector<SynError>& errors,
                         vector<SynElement>& elements) {
     mpreal* r = nullptr;
     int base = 10;
-    if (GetNUmberAt(entry, idx, next_idx, base, &r)) {
+    if (GetNumberAt(entry, idx, next_idx, base, &r)) {
         elements.push_back({kNumber, .re = r, .re_base = base});
         return true;
     } else {
@@ -195,7 +195,7 @@ bool Lexer::ParseNumber(string& entry, size_t idx, size_t& next_idx, vector<SynE
     }
 }
 
-bool Lexer::ParseComplex(string& entry, size_t idx, size_t& next_idx, vector<SynError>& errors __attribute__((unused)),
+bool Lexer::ParseComplex(const string& entry, size_t idx, size_t& next_idx, vector<SynError>& errors __attribute__((unused)),
                          vector<SynElement>& elements) {
     mpreal* re = nullptr;
     mpreal* im = nullptr;
@@ -205,7 +205,7 @@ bool Lexer::ParseComplex(string& entry, size_t idx, size_t& next_idx, vector<Syn
         next_idx = entry.size();
         return true;  // complex format error, return a symbol
     }
-    if (!GetNUmberAt(entry, idx + 1, next_idx, re_base, &re, ',')) {
+    if (!GetNumberAt(entry, idx + 1, next_idx, re_base, &re, ',')) {
         elements.push_back({kSymbol, .value = entry.substr(idx, entry.size() - idx)});
         next_idx = entry.size();
         return true;  // complex format error, return a symbol
@@ -220,7 +220,7 @@ bool Lexer::ParseComplex(string& entry, size_t idx, size_t& next_idx, vector<Syn
         return true;  // complex format error, return a symbol
     }
 
-    if (!GetNUmberAt(entry, i, next_idx, im_base, &im, ')')) {
+    if (!GetNumberAt(entry, i, next_idx, im_base, &im, ')')) {
         elements.push_back({kSymbol, .value = entry.substr(idx, entry.size() - idx)});
         next_idx = entry.size();
         if (re != nullptr) delete re;
@@ -232,7 +232,7 @@ bool Lexer::ParseComplex(string& entry, size_t idx, size_t& next_idx, vector<Syn
     return true;
 }
 
-bool Lexer::ParseReserved(string& entry, size_t idx, size_t& next_idx, vector<SynElement>& elements,
+bool Lexer::ParseReserved(const string& entry, size_t idx, size_t& next_idx, vector<SynElement>& elements,
                           map<string, ReservedWord>& keywords) {
     stringstream ss(entry.substr(idx));
     string token;
@@ -247,7 +247,7 @@ bool Lexer::ParseReserved(string& entry, size_t idx, size_t& next_idx, vector<Sy
     return false;
 }
 
-bool Lexer::ParseUnknown(string& entry, size_t idx, size_t& next_idx, vector<SynElement>& elements) {
+bool Lexer::ParseUnknown(const string& entry, size_t idx, size_t& next_idx, vector<SynElement>& elements) {
     stringstream ss(entry.substr(idx));
     string token;
     ss >> token;