## Nested programs parsing gain
Programs are parsed completely at rt.

Before:
- Entry `<< << << <<` parsed to program `«<< << <<»`
- Entry `<< -> n << n 2 * >> >>` parsed to program `«-> n << n 2 * >>»`

After:
- Entry `<< << << <<` parsed to program `« «  » »`
- Entry `<< -> n << n 2 * >> >>` parsed to program `« -> n « n 2 * » »`

## Performance gain
After tests on fibo `<< dup 1 > if then dup 1 - fibo swap 2 - fibo + else 1 == 1 0 ifte end >> 'fibo' sto 24 fibo`
Super ! gain is 1 to 8 !
```
rpn version is v2.4-11-gd705cf2

Total: run 0 tests: 0 passed, 0 failed (0 steps: 0 passed, 0 failed)

real	0m3,197s
user	0m3,187s
sys	0m0,001s

rpn version is v2.4-12-gac7903b

Total: run 0 tests: 0 passed, 0 failed (0 steps: 0 passed, 0 failed)

real	0m0,348s
user	0m0,345s
sys	0m0,001s
```